+++ changelog.md
@@ -10,6 +10,7 @@
 - Added option to show a label for the file format for each file
 - Added new hooks
 - Added hint which page and view in backend is added or extended by this plugin
+- Added global settings for our block which can also be inherited to all blocks
 
 ### Changed